Add CustomException EmployeeController

+++ App/library-management/src/main/java/com/arobs/internship/librarymanagement/repository/jdbc/mapper/TagMapper.java
@@ -11,7 +11,7 @@ public class TagMapper implements RowMapper<Tag> {
     @Override
     public Tag mapRow(ResultSet resultSet, int i) throws SQLException {
         Tag tag = new Tag();
-        tag.setId(resultSet.getLong("tag_id"));
+        tag.setId(resultSet.getInt("tag_id"));
         tag.setTagName(resultSet.getString("tag_name"));
 
         return tag;
